refactor(init): import PackageManager from lib directly
Remove the PackageManager type re-export from bootstrap-registry and
update consumers (bootstrap.ts, context.ts, index.ts) to import the
type directly from lib/package-manager.ts.
